+++ client/src/components/Header/components/activities/Activities.jsx
@@ -4,10 +4,21 @@ import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
 import SearchIcon from '@mui/icons-material/Search';
 import LoginIcon from '@mui/icons-material/Login';
 import './style.scss';
+import Cart from "../../../Cart/Cart";
 
 function Activities() {
+
+	const [open, setOpen] = React.useState(false);
+	const [scroll, setScroll] = React.useState('paper');
+
+	const handleClickOpen = () => () => {
+		setOpen(true);
+		setScroll('paper');
+	};
+
 	return (
 		<>
+			<Cart open={open} setOpen={setOpen} scroll={scroll} setScroll={setScroll} handleClickOpen={handleClickOpen}/>
 			<Grid item container lg={7} sm={7} alignItems="center">
 				<Typography className="header__links" component="a" href="/">
 					Reviews
@@ -17,7 +28,7 @@ function Activities() {
 				</Typography>
 			</Grid>
 			<Grid container item justifyContent="space-evenly" sm={4}>
-				<Button className="header__buttons" disableRipple size="small">
+				<Button className="header__buttons" disableRipple size="small" onClick={handleClickOpen('paper')}>
 					<ShoppingCartIcon
 						color="grayColor"
 						sx={{
